+++ packages/components/buttons/flat-button/index.js
@@ -1,3 +1,3 @@
-export { default } from '../../../../src/components/buttons/flat-button';
+export { default } from './src';
 
 export { default as version } from './version';
