+++ packages/components/buttons/flat-button/src/index.js
