+++ src/components/field-label/field-label.story.js
@@ -2,8 +2,8 @@ import React from 'react';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 import { withKnobs, boolean, text, select } from '@storybook/addon-knobs/react';
+import FlatButton from '@commercetools-uikit/flat-button';
 import * as icons from '../icons';
-import FlatButton from '../buttons/flat-button';
 import Section from '../../../.storybook/decorators/section';
 import Readme from './README.md';
 import FieldLabel from './field-label';
